Revert "Adding emitpy silicon tests for data movement ops (#5505)"
This reverts commit 63b752c80b71d1e07ff5b3619440fea46b57b0f4.
